+++ gaios/Views/SquareSliderView/SquareSliderView.swift
@@ -11,6 +11,9 @@ class SquareSliderView: UIView {
     var slideLbl = UILabel()
     var slideThumb = UIView()
 
+    var slideStr = "id_slide_to_send".localized
+    var sendingStr = "id_sending".localized
+
     func commonInit() {
         self.translatesAutoresizingMaskIntoConstraints = false
         let pan = UIPanGestureRecognizer(target: self, action: #selector(onPan(_:)))
@@ -48,7 +51,7 @@ class SquareSliderView: UIView {
                 }, completion: { _ in
                     self.delegate?.sliderThumbDidStopMoving(1)
                     self.slideLbl.alpha = 1.0
-                    self.slideLbl.text = NSLocalizedString("id_sending", comment: "")
+                    self.slideLbl.text = self.sendingStr
                 })
             } else {
                 UIView.animate(withDuration: 0.15, animations: {
@@ -56,7 +59,7 @@ class SquareSliderView: UIView {
                 }, completion: { _ in
                     self.delegate?.sliderThumbDidStopMoving(0)
                     self.slideLbl.alpha = 1.0
-                    self.slideLbl.text = NSLocalizedString("id_slide_to_send", comment: "")
+                    self.slideLbl.text = self.slideStr
                 })
             }
         default:
@@ -70,7 +73,6 @@ class SquareSliderView: UIView {
         let fH: CGFloat = self.frame.size.height
 
         self.subviews.forEach { e in e.removeFromSuperview() }
-        let bgH = 0.9
 
         let bg = UIView(frame: CGRect(x: 0.0, y: 0.0, width: self.frame.size.width, height: self.frame.size.height))
         bg.backgroundColor = .black
@@ -80,7 +82,7 @@ class SquareSliderView: UIView {
         self.slideLbl = UILabel(frame: CGRect(x: 0.0, y: 0.0, width: self.frame.size.width, height: self.frame.size.height))
         self.slideLbl.textAlignment = .center
         self.slideLbl.font = .systemFont(ofSize: 16.0, weight: .semibold)
-        self.slideLbl.text = NSLocalizedString("id_slide_to_send", comment: "")
+        self.slideLbl.text = self.slideStr
         self.slideLbl.textColor = .white
         self.addSubview(self.slideLbl)
 
@@ -102,7 +104,12 @@ class SquareSliderView: UIView {
         }, completion: { _ in
             self.delegate?.sliderThumbDidStopMoving(0)
             self.slideLbl.alpha = 1.0
-            self.slideLbl.text = NSLocalizedString("id_slide_to_send", comment: "")
+            self.slideLbl.text = self.slideStr
         })
+    }
+
+    func isActive(_ isActive: Bool) {
+        isUserInteractionEnabled = isActive
+        self.slideThumb.backgroundColor = isActive ? UIColor.gGreenMatrix() : UIColor.gW40()
     }
 }